backend/pyproject.toml
[tool.poetry]
name = "your-virtual-env"
version = "0.1.0"
description = "FSI ThreatSight360"
authors = ["Mehar Grewal <mehar.grewal@mongodb.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
boto3 = "^1.35.70"
botocore = "^1.35.70"
motor = "^3.7.0"
websockets = "^15.0.1"
azure-ai-inference = "*"
azure-identity = "*"
azure-core = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
